Move all hooks to shared hooks folder
