--- frontend/src/hooks/useAudio.ts
@@ -1,7 +1,0 @@
-function useAudio(audioPath: string) {
-	const audio = new Audio(audioPath);
-
-	return audio;
-}
-
-export default useAudio
