+++ frontend/src/shared/hooks/useAudio.ts
@@ -1,0 +1,7 @@
+function useAudio(audioPath: string) {
+	const audio = new Audio(audioPath);
+
+	return audio;
+}
+
+export default useAudio
