+++ frontend/src/shared/hooks/useWebsocketWithLifecycle.ts
